requirements.txt
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.138
botocore==1.34.138
cachetools==5.3.3
cement==2.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.3
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.6
django-ckeditor-5==0.2.13
django-compat==1.0.15
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
django-oauth-toolkit==2.4.0
django-rest-passwordreset==1.4.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-extra-fields==3.7.0
drf-social-oauth2==2.2.0
drf-yasg==1.21.7
filetype==1.2.0
google-auth==2.31.0
google-auth-oauthlib==1.2.0
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
mailchimp-marketing==3.0.80
mailchimp-transactional==1.0.56
mailjet-rest==1.3.4
numpy==1.26.4
oauth2-provider==0.0
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pathspec==0.10.1
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
semantic-version==2.8.5
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
stripe==9.10.0
termcolor==1.1.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.1.9
whitenoise==6.6.0
